유저없으면500에러말고null반환

+++ src/main/java/invincibleDevs/bookpago/review/MyReviewDto.java
@@ -1,17 +1,14 @@
 package invincibleDevs.bookpago.review;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-
 public record MyReviewDto(
         String bookImageUrl,
-     String bookTitle,
-     String bookAuthor,
-     Long id,
-     String content,
-     Long isbn,
-     int rating,
-     Long authorId
+        String bookTitle,
+        String bookAuthor,
+        Long id,
+        String content,
+        Long isbn,
+        double rating,
+        Long authorId
 ) {
+
 }
